+++ packages/dx-grid-core/src/plugins/local-grouping/computeds.test.js
@@ -31,6 +31,13 @@ describe('LocalGrouping computeds', () => {
     });
   });
 
+  const groupRow = ({ groupedBy, ...restParams }) => ({
+    ...restParams,
+    groupedBy,
+    [GRID_GROUP_CHECK]: true,
+    [GRID_GROUP_LEVEL_KEY]: `${GRID_GROUP_TYPE}_${groupedBy}`,
+  });
+
   const rows = [
     { a: 1, b: 1 },
     { a: 1, b: 2 },
@@ -40,56 +47,68 @@ describe('LocalGrouping computeds', () => {
   const getCellValue = (row, columnName) => row[columnName];
 
   const firstGrouping = [{ columnName: 'a' }];
-  const firstGroupedRows = [{
-    value: 1,
-    key: '1',
-    items: [
-      { a: 1, b: 1 },
-      { a: 1, b: 2 },
-    ],
-  }, {
-    value: 2,
-    key: '2',
-    items: [
-      { a: 2, b: 1 },
-      { a: 2, b: 2 },
-    ],
-  }];
+  const firstGroupedRows = [
+    groupRow({
+      groupedBy: 'a',
+      compoundKey: '1',
+      key: '1',
+      value: 1,
+    }),
+    { a: 1, b: 1 },
+    { a: 1, b: 2 },
+    groupRow({
+      groupedBy: 'a',
+      compoundKey: '2',
+      key: '2',
+      value: 2,
+    }),
+    { a: 2, b: 1 },
+    { a: 2, b: 2 },
+  ];
 
   const secondGrouping = [{ columnName: 'a' }, { columnName: 'b' }];
-  const secondGroupedRows = [{
-    value: 1,
-    key: '1',
-    items: [{
-      value: 1,
+  const secondGroupedRows = [
+    groupRow({
+      groupedBy: 'a',
+      compoundKey: '1',
       key: '1',
-      items: [
-        { a: 1, b: 1 },
-      ],
-    }, {
-      value: 2,
-      key: '2',
-      items: [
-        { a: 1, b: 2 },
-      ],
-    }],
-  }, {
-    value: 2,
-    key: '2',
-    items: [{
       value: 1,
+    }),
+    groupRow({
+      groupedBy: 'b',
+      compoundKey: '1|1',
       key: '1',
-      items: [
-        { a: 2, b: 1 },
-      ],
-    }, {
-      value: 2,
+      value: 1,
+    }),
+    { a: 1, b: 1 },
+    groupRow({
+      groupedBy: 'b',
+      compoundKey: '1|2',
       key: '2',
-      items: [
-        { a: 2, b: 2 },
-      ],
-    }],
-  }];
+      value: 2,
+    }),
+    { a: 1, b: 2 },
+    groupRow({
+      groupedBy: 'a',
+      compoundKey: '2',
+      key: '2',
+      value: 2,
+    }),
+    groupRow({
+      groupedBy: 'b',
+      compoundKey: '2|1',
+      key: '1',
+      value: 1,
+    }),
+    { a: 2, b: 1 },
+    groupRow({
+      groupedBy: 'b',
+      compoundKey: '2|2',
+      key: '2',
+      value: 2,
+    }),
+    { a: 2, b: 2 },
+  ];
 
   describe('#groupedRows', () => {
     it('can group by first column', () => {
@@ -108,21 +127,24 @@ describe('LocalGrouping computeds', () => {
         value: `${value}_test`,
       });
       expect(groupedRows(rows, firstGrouping, getCellValue, getColumnIdentity))
-        .toEqual([{
-          value: '1_test',
-          key: '1',
-          items: [
-            { a: 1, b: 1 },
-            { a: 1, b: 2 },
-          ],
-        }, {
-          value: '2_test',
-          key: '2',
-          items: [
-            { a: 2, b: 1 },
-            { a: 2, b: 2 },
-          ],
-        }]);
+        .toEqual([
+          groupRow({
+            groupedBy: 'a',
+            compoundKey: '1',
+            key: '1',
+            value: '1_test',
+          }),
+          { a: 1, b: 1 },
+          { a: 1, b: 2 },
+          groupRow({
+            groupedBy: 'a',
+            compoundKey: '2',
+            key: '2',
+            value: '2_test',
+          }),
+          { a: 2, b: 1 },
+          { a: 2, b: 2 },
+        ]);
     });
 
     it('should use getColumnIdentity argument for each grouping', () => {
@@ -131,39 +153,48 @@ describe('LocalGrouping computeds', () => {
       });
 
       expect(groupedRows(rows, secondGrouping, getCellValue, getColumnIdentity))
-        .toEqual([{
-          value: '1_test',
-          key: '1_test',
-          items: [{
-            value: '1_test',
+        .toEqual([
+          groupRow({
+            groupedBy: 'a',
+            compoundKey: '1_test',
             key: '1_test',
-            items: [
-              { a: 1, b: 1 },
-            ],
-          }, {
-            value: '2_test',
-            key: '2_test',
-            items: [
-              { a: 1, b: 2 },
-            ],
-          }],
-        }, {
-          value: '2_test',
-          key: '2_test',
-          items: [{
             value: '1_test',
+          }),
+          groupRow({
+            groupedBy: 'b',
+            compoundKey: '1_test|1_test',
             key: '1_test',
-            items: [
-              { a: 2, b: 1 },
-            ],
-          }, {
-            value: '2_test',
+            value: '1_test',
+          }),
+          { a: 1, b: 1 },
+          groupRow({
+            groupedBy: 'b',
+            compoundKey: '1_test|2_test',
             key: '2_test',
-            items: [
-              { a: 2, b: 2 },
-            ],
-          }],
-        }]);
+            value: '2_test',
+          }),
+          { a: 1, b: 2 },
+          groupRow({
+            groupedBy: 'a',
+            compoundKey: '2_test',
+            key: '2_test',
+            value: '2_test',
+          }),
+          groupRow({
+            groupedBy: 'b',
+            compoundKey: '2_test|1_test',
+            key: '1_test',
+            value: '1_test',
+          }),
+          { a: 2, b: 1 },
+          groupRow({
+            groupedBy: 'b',
+            compoundKey: '2_test|2_test',
+            key: '2_test',
+            value: '2_test',
+          }),
+          { a: 2, b: 2 },
+        ]);
     });
 
     it('should pass column name to getColumnIdentity', () => {
@@ -188,22 +219,20 @@ describe('LocalGrouping computeds', () => {
 
       expect(expandedGroupRows(firstGroupedRows, firstGrouping, expandedGroups))
         .toEqual([
-          {
-            [GRID_GROUP_CHECK]: true,
-            [GRID_GROUP_LEVEL_KEY]: `${GRID_GROUP_TYPE}_a`,
+          groupRow({
             groupedBy: 'a',
+            compoundKey: '1',
             key: '1',
             value: 1,
-          },
+          }),
           { a: 1, b: 1 },
           { a: 1, b: 2 },
-          {
-            [GRID_GROUP_CHECK]: true,
-            [GRID_GROUP_LEVEL_KEY]: `${GRID_GROUP_TYPE}_a`,
+          groupRow({
             groupedBy: 'a',
+            compoundKey: '2',
             key: '2',
             value: 2,
-          },
+          }),
         ]);
     });
 
@@ -212,35 +241,31 @@ describe('LocalGrouping computeds', () => {
 
       expect(expandedGroupRows(secondGroupedRows, secondGrouping, expandedGroups))
         .toEqual([
-          {
-            [GRID_GROUP_CHECK]: true,
-            [GRID_GROUP_LEVEL_KEY]: `${GRID_GROUP_TYPE}_a`,
+          groupRow({
             groupedBy: 'a',
+            compoundKey: '1',
             key: '1',
             value: 1,
-          },
-          {
-            [GRID_GROUP_CHECK]: true,
-            [GRID_GROUP_LEVEL_KEY]: `${GRID_GROUP_TYPE}_b`,
+          }),
+          groupRow({
             groupedBy: 'b',
-            key: '1|1',
+            compoundKey: '1|1',
+            key: '1',
             value: 1,
-          },
-          {
-            [GRID_GROUP_CHECK]: true,
-            [GRID_GROUP_LEVEL_KEY]: `${GRID_GROUP_TYPE}_b`,
+          }),
+          groupRow({
             groupedBy: 'b',
-            key: '1|2',
-            value: 2,
-          },
-          { a: 1, b: 2 },
-          {
-            [GRID_GROUP_CHECK]: true,
-            [GRID_GROUP_LEVEL_KEY]: `${GRID_GROUP_TYPE}_a`,
-            groupedBy: 'a',
+            compoundKey: '1|2',
             key: '2',
             value: 2,
-          },
+          }),
+          { a: 1, b: 2 },
+          groupRow({
+            groupedBy: 'a',
+            compoundKey: '2',
+            key: '2',
+            value: 2,
+          }),
         ]);
     });
   });