+++ packages/dx-grid-core/src/plugins/selection-state/computeds.js
@@ -1,6 +1,10 @@
-export const getAvailableToSelect = (rows, getRowId, isGroupRow) => rows
-  .filter(row => (!isGroupRow || !isGroupRow(row)) && !row.type)
-  .map(row => getRowId(row));
+export const getAvailableToSelect = (rows, getRowId, isGroupRow) => {
+  let dataRows = rows;
+  if (isGroupRow) {
+    dataRows = dataRows.filter(row => !isGroupRow(row));
+  }
+  return dataRows.map(row => getRowId(row));
+};
 
 export const getAvailableSelection = (selection, availableToSelect) => {
   const availableToSelectSet = new Set(availableToSelect);