+++ packages/dx-react-demos/src/material-ui/featured-redux-integration/demo.jsx
@@ -5,12 +5,12 @@ import { connect, Provider } from 'react-redux';
 import {
   SortingState, SelectionState, FilteringState, PagingState, GroupingState, RowDetailState,
   LocalFiltering, LocalGrouping, LocalPaging, LocalSorting,
-  ColumnOrderState, TableColumnResizing,
+  ColumnOrderState,
 } from '@devexpress/dx-react-grid';
 import {
   Grid,
   TableView, TableHeaderRow, TableFilterRow, TableSelection, TableGroupRow, TableRowDetail,
-  GroupingPanel, PagingPanel, DragDropContext,
+  GroupingPanel, PagingPanel, DragDropContext, TableColumnResizing,
 } from '@devexpress/dx-react-grid-material-ui';
 
 import { withStyles } from 'material-ui/styles';
@@ -268,6 +268,7 @@ const ReduxGridContainer = connect(mapStateToProps, mapDispatchToProps)(GridCont
 const store = createStore(
   gridReducer,
   // Enabling Redux DevTools Extension (https://github.com/zalmoxisus/redux-devtools-extension)
+  // eslint-disable-next-line no-underscore-dangle
   window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__(),
 );