+++ packages/dx-react-demos/src/material-ui/grouping/local-grouping-custom.test.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { mount } from 'enzyme';
 import Demo from './local-grouping-custom';
 
-describe('MUI: grouping with custom', () => {
+describe('MUI: local-grouping-custom', () => {
   it('should work', () => {
     mount(<Demo />);
   });